# swagger/pelican-swagger.yaml
swagger: "2.0"
info:
  title: Pelican Server APIs
  description:
    "[Pelican](https://pelicanplatform.org/) provides an open-source software platform for federating
    dataset repositories together and delivering the objects to computing capacity such as the [OSPool](https://osg-htc.org/services/open_science_pool.html)


    This is the API documentation for various APIs in Pelican servers (director, registry, origin, etc)
    to communicate with each other and in-between users accessing the servers.


    Note that we use cookie authentication and authorization. We check a cookie named `login` with value being a JWT.
    The cookie is issued after a successful call to `/api/v1.0/auth/login`. However, OpenAPI 2.0 does not support specifying cookie-based security check.
    Therefore, as an alternative, we will add `Authentication Required` to the API description where needed.

    For endpoint that is only accessible by users with admin privilege, we mark the endpoint as `Admin privilege Required`.


    For how to set up Pelican servers, please refer to the documentation at [docs.pelicanplatform.org](https://docs.pelicanplatform.org/)"
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: API Support via Pelican GitHub Issue
    url: https://github.com/PelicanPlatform/pelican/issues
  version: "1.0"
basePath: /api/v1.0/
consumes:
  - application/json
produces:
  - application/json
schemes:
  - https
securityDefinitions:
  # This is a hacky way to specify Bearer token auth here because it's not actually an "apikey"
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >-
      Enter the JWT with the `Bearer` prefix, e.g. "Bearer abcde12345".
definitions:
...
  ErrorModelV2:
    type: object
    description: The error response of a request
    properties:
      status:
        type: string
        default: "error"
        description: The response status
      msg:
        type: string
        default: ""
        description: The response message
...
  Group:
    type: object
    description: >-
      A group is a set of users; membership is one of the inputs to
      authorization. The four-concept user/group model (see
      `docs/user-group-design.md`) gives every group BOTH a
      machine-readable `name` (used in policy strings; admin-controlled)
      and a human-readable `displayName` (UI label; owner-editable). Any
      authorization-affecting UI should render BOTH.
    properties:
      id:
        type: string
        description: Opaque internal primary key. Auto-generated, immutable, never used for authorization.
        example: "abc12345"
      name:
        type: string
        description: >-
          Machine-readable identifier. Used in admin lists, ACL grants,
          and policy strings. Validated by ValidateIdentifier (allowed
          characters A-Za-z0-9._@-, length 2-64; '/' explicitly banned).
          **Editable only by system administrators.** Group owners and
          group admins cannot rename a group.
        example: "research-team"
      displayName:
        type: string
        description: >-
          Human-readable label rendered in the UI. Owner-editable. UIs
          that don't have a display name should fall back to `name`.
        example: "Climate Research Team"
      description:
        type: string
        description: Free-form description shown in the UI.
        example: "Team working on climate research"
      createdBy:
        type: string
        description: >-
          User ID of whoever caused this record to exist. May be a real
          user ID or the sentinel "unknown" for rows that predate the
          audit fields.
        example: "abc12345"
      creatorAuthMethod:
        type: string
        description: How the creator was authenticated when minting the group.
        enum:
          - ""
          - web-cookie
          - api-token
          - bearer-jwt
        example: "web-cookie"
      creatorAuthMethodId:
        type: string
        description: When creatorAuthMethod is "api-token", the 5-character API-token ID prefix; empty otherwise.
        example: "AbCdE"
      ownerId:
        type: string
        description: User ID of the group owner (can transfer ownership, set admin, manage members).
        example: "abc12345"
      adminId:
        type: string
        description: User or group ID delegated as group administrator (can manage members).
        example: "def67890"
      adminType:
        type: string
        description: Whether adminId refers to a 'user' or a 'group'.
        enum:
          - ""
          - user
          - group
        example: "user"
      createdAt:
        type: string
        format: date-time
        example: "2024-01-15T10:30:00Z"
      updatedAt:
        type: string
        format: date-time
      members:
        type: array
        items:
          $ref: "#/definitions/GroupMember"
        description: Group members (typically only populated by GET /groups/{id}, not by list endpoints).
      ownerUser:
        $ref: "#/definitions/UserCard"
      adminUser:
        $ref: "#/definitions/UserCard"
      adminGroup:
        $ref: "#/definitions/GroupCard"
      createdByUser:
        $ref: "#/definitions/UserCard"
  User:
    type: object
    description: >-
      Canonical user record. Per the four-concept model
      (`docs/user-group-design.md`): `id` is opaque-internal,
      `username` is the authorization handle (admin-controlled),
      `displayName` is a human label (self-editable), and (`sub`,
      `issuer`) is the *primary* linked OIDC identity (login linkage
      only — never used for authorization). Additional identities live
      in user_identities.
    properties:
      id:
        type: string
        description: Opaque internal primary key. Immutable; never reused (deletes are soft).
      username:
        type: string
        description: >-
          Authorization handle. Validated by ValidateIdentifier.
          Admin-only renames after first creation; on rename, if the
          user is a local-issuer account, the primary `sub` is kept in
          lockstep (otherwise password login would silently break).
      sub:
        type: string
        description: >-
          The 'sub' claim of the user's *primary* linked OIDC identity.
          For local-issuer users this equals `username`. Used for login
          linkage; never used for authorization.
      issuer:
        type: string
        description: The 'iss' claim of the user's primary linked OIDC identity.
      status:
        type: string
        description: User account status.
        enum:
          - active
          - inactive
        example: "active"
      lastLoginAt:
        type: string
        format: date-time
        description: Timestamp of the user's last login.
      displayName:
        type: string
        description: Human-readable label. Self-editable via PATCH /me. Refreshed from the IdP on each OIDC login.
        example: "John Doe"
      aupVersion:
        type: string
        description: Version of the AUP the user agreed to.
      aupAgreedAt:
        type: string
        format: date-time
        description: Timestamp when the user agreed to the AUP.
      createdBy:
        type: string
        description: >-
          User ID of whoever caused this record to exist, or one of the
          sentinels "self-enrolled" (auto-created on first OIDC sign-in
          or htpasswd login) or "unknown" (pre-audit-column rows).
      creatorAuthMethod:
        type: string
        description: How the creator was authenticated when the row was inserted.
        enum:
          - ""
          - web-cookie
          - api-token
          - bearer-jwt
      creatorAuthMethodId:
        type: string
        description: When creatorAuthMethod is "api-token", the 5-character API-token ID prefix.
      createdAt:
        type: string
        format: date-time
      updatedAt:
        type: string
        format: date-time
  UserCard:
    type: object
    description: >-
      Public-safe user summary used to render "Display Name (username)"
      labels in places that don't need the full record and shouldn't
      require user-listing privilege.
    properties:
      id:
        type: string
      username:
        type: string
      displayName:
        type: string
  GroupCard:
    type: object
    description: Public-safe group summary, analogous to UserCard.
    properties:
      id:
        type: string
      name:
        type: string
  GroupMember:
    type: object
    properties:
      groupId:
        type: string
        description: ID of the group
        example: "abc12345"
      userId:
        type: string
        description: ID of the user
      user:
        $ref: "#/definitions/User"
      createdBy:
        type: string
        description: User who added this member
        example: "admin@example.com"
      createdAt:
        type: string
        format: date-time
        description: When the member was added
        example: "2024-01-15T10:30:00Z"
...
  CreateGroupRequest:
    type: object
    required:
      - name
    properties:
      name:
        type: string
        description: >-
          Machine-readable group identifier. Validated by ValidateIdentifier
          (allowed characters A-Za-z0-9._@-, length 2-64; '/' explicitly
          banned). Cannot start with the reserved 'user-' prefix.
        example: "research-team"
      displayName:
        type: string
        description: Human-readable label shown in the UI. Optional; falls back to `name` when empty.
        example: "Climate Research Team"
      description:
        type: string
        description: Description of the group.
        example: "Team working on climate research"
...
paths:
...
  /groups:
    get:
      tags:
        - groups
      summary: List all groups
      description: "`Authentication Required`"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: array
            items:
              $ref: "#/definitions/Group"
        "401":
          description: Authentication required
          schema:
            $ref: "#/definitions/ErrorModelV2"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorModelV2"
    post:
      tags:
        - groups
      summary: Create a new group
      description: "`Authentication Required`"
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: group
          description: The group data to create
          required: true
          schema:
            $ref: "#/definitions/CreateGroupRequest"
      responses:
        "201":
          description: Group created successfully
          schema:
            $ref: "#/definitions/Group"
        "400":
          description: Bad request
          schema:
            $ref: "#/definitions/ErrorModelV2"
        "401":
          description: Authentication required
          schema:
            $ref: "#/definitions/ErrorModelV2"
        "500":
          description: Internal server error
          schema:
            $ref: "#/definitions/ErrorModelV2"
...